Fix and improve stress tests
